# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/transactions": {
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all transactions and invoices",
        "tags": [
          "Transactions"
        ]
      },
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/payments": {
      "get": {
        "operationId": "TransactionsController_findTransactions",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List transactions only",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/invoices": {
      "get": {
        "operationId": "TransactionsController_findInvoices",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List invoices only",
        "tags": [
          "Transactions"
        ]
      },
      "post": {
        "operationId": "TransactionsController_createInvoice",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create an invoice",
        "tags": [
          "Transactions"
        ]
      }
    },
...
    "/api/transactions/payments/{transactionId}/status": {
      "patch": {
        "operationId": "TransactionsController_updateTransactionStatus",
        "parameters": [
          {
            "name": "transactionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Approve, reject, or update submitted payment transaction status",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/settings/platform": {
      "get": {
        "operationId": "SettingsController_getPlatform",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get platform settings",
        "tags": [
          "Settings"
        ]
      },
      "put": {
        "operationId": "SettingsController_updatePlatform",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlatformSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update platform settings",
        "tags": [
          "Settings"
        ]
      }
    },
    "/api/settings/security": {
      "get": {
        "operationId": "SettingsController_getSecurity",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get security settings",
        "tags": [
          "Settings"
        ]
      },
      "put": {
        "operationId": "SettingsController_updateSecurity",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSecuritySettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update security settings",
        "tags": [
          "Settings"
        ]
      }
    },
    "/api/settings/permissions": {
      "get": {
        "operationId": "SettingsController_getAllPermissions",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all role permissions",
        "tags": [
          "Settings"
        ]
      }
    },
    "/api/settings/permissions/{role}": {
      "get": {
        "operationId": "SettingsController_getPermissions",
        "parameters": [
          {
            "name": "role",
            "required": true,
            "in": "path",
            "description": "Role key (fleet-manager, business-client, driver)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get permissions for a specific role",
        "tags": [
          "Settings"
        ]
      },
      "put": {
        "operationId": "SettingsController_updatePermissions",
        "parameters": [
          {
            "name": "role",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update permissions for a specific role",
        "tags": [
          "Settings"
        ]
      }
    },
...
  },
  "info": {
    "title": "DeliverSync API",
    "description": "REST API for the DeliverSync delivery management platform. Role-based access is controlled via the `x-user-role` request header. Supported roles: superuser, fleet-manager, business-client, driver.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePlatformSettingsDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "DeliverSync"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata"
          },
          "language": {
            "type": "string",
            "example": "English"
          },
          "logo": {
            "type": "string",
            "example": ""
          }
        }
      },
      "UpdateSecuritySettingsDto": {
        "type": "object",
        "properties": {
          "passwordLength": {
            "type": "number",
            "example": 8
          },
          "failedAttempts": {
            "type": "number",
            "example": 5
          },
          "sessionTimeout": {
            "type": "number",
            "example": 30
          },
          "twoFactor": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "UpdatePermissionsDto": {
        "type": "object",
        "properties": {
          "permissions": {
            "description": "Array of permission entries: [label, description, enabled]",
            "example": [
              [
                "Dashboard Access",
                "View platform overview",
                true
              ]
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "permissions"
        ]
      }
    }
  }
}